transcoding: move transcoding to its own folder
