+++ src/libav.c
@@ -1,4 +1,4 @@
-#include "plumbing/transcoding.h"
+#include "transcoding/transcoding.h"
 #include "libav.h"
 
 /**
@@ -179,8 +179,8 @@ codec_id2streaming_component_type(enum AVCodecID id)
 
 
 /**
- * 
- */ 
+ *
+ */
 int
 libav_is_encoder(AVCodec *codec)
 {
@@ -192,8 +192,8 @@ libav_is_encoder(AVCodec *codec)
 }
 
 /**
- * 
- */ 
+ *
+ */
 void
 libav_set_loglevel(void)
 {
